fix(test): temporarily move tests from impl to autoconfigure module
