+++ auto-configurations/vector-stores/spring-ai-alibaba-autoconfigure-vector-store-analyticdb/src/test/resources/time.shelter.txt
@@ -1,0 +1,2 @@
+Somewhere in the Andes, they believe in this very day that the future is behind you. It comes up from behind your back, surprising and unforeseeable, while the past is always before your eyes, that which has already happened. When they talk about the past, the people of the Aymara tribe point in front of them. You walk forward facing the past, and you turn back toward the future.
+― Georgi Gospodinov, Time Shelter
