+++ auto-configurations/vector-stores/spring-ai-alibaba-autoconfigure-vector-store-tablestore/src/test/java/com/alibaba/cloud/ai/vectorstore/tablestore/autoconfigure/TestEmbedding.java
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2024-2026 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.cloud.ai.vectorstore.tablestore.autoconfigure;
+
+import dev.langchain4j.model.embedding.onnx.allminilml6v2q.AllMiniLmL6V2QuantizedEmbeddingModel;
+import org.springframework.ai.document.Document;
+import org.springframework.ai.embedding.Embedding;
+import org.springframework.ai.embedding.EmbeddingModel;
+import org.springframework.ai.embedding.EmbeddingRequest;
+import org.springframework.ai.embedding.EmbeddingResponse;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TestEmbedding implements EmbeddingModel {
+
+	private final AllMiniLmL6V2QuantizedEmbeddingModel embeddingModel = new AllMiniLmL6V2QuantizedEmbeddingModel();
+
+	@Override
+	public EmbeddingResponse call(EmbeddingRequest request) {
+		int size = request.getInstructions().size();
+		List<Embedding> embeddings = new ArrayList<>(size);
+		for (int i = 0; i < size; i++) {
+			embeddings.add(new Embedding(embeddingModel.embed(request.getInstructions().get(i)).content().vector(), i));
+		}
+		return new EmbeddingResponse(embeddings);
+	}
+
+	@Override
+	public float[] embed(Document document) {
+		return embeddingModel.embed(document.getText()).content().vector();
+	}
+
+}
